feat(core): use reqq field from extension handshake (#20)
Introduce `reqq` header in extension handshake to improve download
scheduling

This commit adds support for a `reqq` (request queue) field in the
extension handshake,
allowing the client to advertise its maximum number of outstanding
requests. By default,
we assume a limit of 250; if a peer advertises a higher limit, we update
accordingly.

While increasing `reqq` can improve throughput, the effect is bounded by
the piece picker.
Requests are scheduled at the piece level, so if pieces are small (given
16 KiB blocks),
the total number of outgoing requests is naturally constrained by the
number of blocks per piece.
Arbitrarily increasing `reqq` beyond this limit does not yield faster
downloads, as the piece picker
becomes the bottleneck.